Rename option focus hook and move to separate folder
