+++ desktop/packages/mullvad-vpn/src/renderer/components/settings-listbox/components/split-option/SplitOption.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Flex } from '../../../../lib/components';
 import { Listbox } from '../../../../lib/components/listbox';
 import { ListboxOptionProps } from '../../../../lib/components/listbox/components';
-import { useOptions } from '../../../../lib/hooks';
+import { useRovingFocus } from '../../../../lib/hooks';
 import { SplitOptionItem, SplitOptionNavigateButton } from './components';
 
 export type SplitOptionProps<T> = ListboxOptionProps<T>;
@@ -11,7 +11,7 @@ export type SplitOptionProps<T> = ListboxOptionProps<T>;
 function SplitOption<T>({ children, ...props }: SplitOptionProps<T>) {
   const optionsRef = React.useRef<HTMLDivElement>(null);
   const [focusedIndex, setFocusedIndex] = React.useState<number | undefined>(undefined);
-  const { handleKeyboardNavigation, handleBlur, handleFocus } = useOptions({
+  const { handleKeyboardNavigation, handleBlur, handleFocus } = useRovingFocus({
     optionsRef,
     orientation: 'horizontal',
     selector: '[data-split-button="true"]:not([aria-disabled="true"])',
