+++ desktop/packages/mullvad-vpn/src/renderer/lib/hooks/index.ts
@@ -1,5 +1,2 @@
 export * from './use-exclusive-task';
-export * from './use-focus-option-by-index';
-export * from './use-get-initial-focus-index';
-export * from './use-handle-options-keyboard-navigation';
-export * from './use-options';
+export * from './use-roving-focus';
