+++ desktop/packages/mullvad-vpn/src/renderer/lib/hooks/use-roving-focus/hooks/use-handle-options-keyboard-navigation/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './use-focus-option-by-index';
+export * from './use-get-initial-focus-index';
