+++ desktop/packages/mullvad-vpn/src/renderer/lib/hooks/use-roving-focus/hooks/use-handle-options-keyboard-navigation/hooks/use-focus-option-by-index.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { getOptions } from '../utils';
+import { getOptions } from '../../../utils';
 
 export const useFocusOptionByIndex = <T extends HTMLElement>({
   optionsRef,
