+++ desktop/packages/mullvad-vpn/src/renderer/lib/hooks/use-roving-focus/hooks/use-handle-options-keyboard-navigation/hooks/use-get-initial-focus-index.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { getOptions, getSelectedOptionIndex } from '../utils';
+import { getOptions, getSelectedOptionIndex } from '../../../utils';
 
 export const useGetInitialFocusIndex = <T extends HTMLElement>({
   focusedIndex,
