+++ desktop/packages/mullvad-vpn/src/renderer/lib/hooks/use-roving-focus/utils/index.ts
@@ -1,4 +1,5 @@
-export * from './get-initial-option';
+export * from './get-is-option-selected';
 export * from './get-options';
+export * from './get-initial-option';
 export * from './get-selected-option-index';
 export * from './get-selected-option';
